refactor: migrate from layer-based to feature-based architecture
