+++ src/app.ts
@@ -1,0 +1,1 @@
+//Khởi tạo app Express, middleware, routes tổng
