+++ src/server.ts
@@ -1,0 +1,1 @@
+//Chạy server (entry point)
